routes: migrate maintenance + search-admin domain (11 routes)
Move to src/server/routes/maintenance.ts:
- /api/maintenance/orphaned-worktrees, cleanup-worktrees
- /api/maintenance/orphaned-sessions, cleanup-sessions
- /api/maintenance/expired-archives, purge-archives
- /api/search/rebuild, /api/search/stats, /api/search/compact
- /api/maintenance/orphaned-index-rows, cleanup-index-rows

Net: server.ts -266 lines (6085 → 5819). No behaviour change.
