+++ src/server/server.ts
@@ -5813,272 +5813,6 @@ async function handleApiRoute(
 		return;
 	}
 
-	// ─── Maintenance endpoints ──────────────────────────────────────────
-	// These replace the old automatic cleanup-on-startup behavior.
-	// Users can preview orphaned resources and choose to clean them up.
-
-	// GET /api/maintenance/orphaned-worktrees
-	if (url.pathname === "/api/maintenance/orphaned-worktrees" && req.method === "GET") {
-		const allOrphans: Array<{ path: string; branch: string; repoPath: string }> = [];
-		for (const ctx of projectContextManager.all()) {
-			try {
-				const repoPath = ctx.project.rootPath;
-				if (await isGitRepo(repoPath)) {
-					const orphans = await sessionManager.listOrphanedSessionWorktrees(repoPath);
-					for (const o of orphans) {
-						allOrphans.push({ ...o, repoPath });
-					}
-				}
-			} catch { /* best-effort */ }
-		}
-		json({ worktrees: allOrphans });
-		return;
-	}
-
-	// POST /api/maintenance/cleanup-worktrees
-	if (url.pathname === "/api/maintenance/cleanup-worktrees" && req.method === "POST") {
-		const body = await readBody(req);
-		let cleaned = 0;
-		if (body?.worktrees && Array.isArray(body.worktrees)) {
-			// Clean specific worktrees — validate each against registered projects and orphan list
-			const validRepoPaths = new Set([...projectContextManager.all()].map(ctx => ctx.project.rootPath));
-			for (const wt of body.worktrees) {
-				if (wt.path && wt.branch && wt.repoPath) {
-					// Validate repoPath is a registered project
-					if (!validRepoPaths.has(wt.repoPath)) continue;
-					// Validate this worktree is actually orphaned
-					try {
-						const orphans = await sessionManager.listOrphanedSessionWorktrees(wt.repoPath);
-						const isOrphan = orphans.some(o => o.path === wt.path && o.branch === wt.branch);
-						if (!isOrphan) continue;
-					} catch { continue; }
-					try {
-						const { cleanupWorktree } = await import("./skills/git.js");
-						await cleanupWorktree(wt.repoPath, wt.path, wt.branch, true);
-						cleaned++;
-					} catch { /* best-effort */ }
-				}
-			}
-		} else {
-			// Clean all orphans across all projects
-			for (const ctx of projectContextManager.all()) {
-				try {
-					const repoPath = ctx.project.rootPath;
-					if (await isGitRepo(repoPath)) {
-						await sessionManager.cleanupOrphanedSessionWorktrees(repoPath);
-						cleaned++; // count projects cleaned, not individual worktrees
-					}
-				} catch { /* best-effort */ }
-			}
-		}
-		json({ cleaned });
-		return;
-	}
-
-	// GET /api/maintenance/orphaned-sessions
-	if (url.pathname === "/api/maintenance/orphaned-sessions" && req.method === "GET") {
-		const sessions = await sessionManager.listOrphanedNonInteractiveSessions();
-		json({ sessions });
-		return;
-	}
-
-	// POST /api/maintenance/cleanup-sessions
-	if (url.pathname === "/api/maintenance/cleanup-sessions" && req.method === "POST") {
-		const body = await readBody(req);
-		const orphans = await sessionManager.listOrphanedNonInteractiveSessions();
-		const orphanIds = new Set(orphans.map(o => o.id));
-		const idsToTerminate = (body?.sessionIds && Array.isArray(body.sessionIds))
-			? (body.sessionIds as string[]).filter(id => orphanIds.has(id))
-			: orphans.map(o => o.id);
-		const terminated = await sessionManager.terminateOrphanedSessions(idsToTerminate);
-		json({ terminated });
-		return;
-	}
-
-	// GET /api/maintenance/expired-archives
-	if (url.pathname === "/api/maintenance/expired-archives" && req.method === "GET") {
-		const stats = await sessionManager.getExpiredArchiveStats();
-		json(stats);
-		return;
-	}
-
-	// POST /api/maintenance/purge-archives
-	if (url.pathname === "/api/maintenance/purge-archives" && req.method === "POST") {
-		await sessionManager.purgeExpiredArchives();
-		const stats = await sessionManager.getExpiredArchiveStats();
-		json({ purged: true, remaining: stats });
-		return;
-	}
-
-	// ─── Search admin endpoints (design §9, §11) ─────────────────────
-
-	function resolveSearchProject(pid: string | undefined | null) {
-		if (!pid) return null;
-		const ctx = projectContextManager.getOrCreate(pid);
-		return ctx;
-	}
-
-	function searchUnavailableResponse(state: string) {
-		const reasonMap: Record<string, string> = {
-			"disabled": "disabled",
-			"closed": "closed",
-			"initializing": "initializing",
-		};
-		const reason = reasonMap[state] ?? state;
-		return { error: "search-unavailable", reason, state };
-	}
-
-	// POST /api/search/rebuild
-	if (url.pathname === "/api/search/rebuild" && req.method === "POST") {
-		const body = await readBody(req);
-		const projectId = body && typeof body === "object" ? (body as any).projectId : undefined;
-		if (!projectId || typeof projectId !== "string") {
-			json({ error: "Missing projectId" }, 400);
-			return;
-		}
-		const ctx = resolveSearchProject(projectId);
-		if (!ctx) { json({ error: "Project not found" }, 404); return; }
-		await ctx.searchIndex.whenReady();
-		const state = ctx.searchIndex.getState();
-		if (state !== "ready") {
-			json(searchUnavailableResponse(state), 503);
-			return;
-		}
-		// Kick off in background — client observes progress over WS.
-		ctx.searchIndex
-			.rebuildFromStores(ctx.goalStore, ctx.sessionStore, undefined, ctx.staffStore)
-			.catch((err) => console.error("[search] rebuild failed:", err));
-		json({ ok: true }, 202);
-		return;
-	}
-
-	// GET /api/search/stats?projectId=...
-	if (url.pathname === "/api/search/stats" && req.method === "GET") {
-		const projectId = url.searchParams.get("projectId") || undefined;
-		if (!projectId) { json({ error: "Missing projectId" }, 400); return; }
-		const ctx = resolveSearchProject(projectId);
-		if (!ctx) { json({ error: "Project not found" }, 404); return; }
-		await ctx.searchIndex.whenReady();
-		const stats = await ctx.searchIndex.getStats();
-		json(stats);
-		return;
-	}
-
-	// POST /api/search/compact
-	if (url.pathname === "/api/search/compact" && req.method === "POST") {
-		const body = await readBody(req);
-		const projectId = body && typeof body === "object" ? (body as any).projectId : undefined;
-		if (!projectId || typeof projectId !== "string") {
-			json({ error: "Missing projectId" }, 400);
-			return;
-		}
-		const ctx = resolveSearchProject(projectId);
-		if (!ctx) { json({ error: "Project not found" }, 404); return; }
-		await ctx.searchIndex.whenReady();
-		const state = ctx.searchIndex.getState();
-		if (state !== "ready") {
-			json(searchUnavailableResponse(state), 503);
-			return;
-		}
-		try {
-			await ctx.searchIndex.compact();
-			json({ ok: true });
-		} catch (err) {
-			json({ error: `Compact failed: ${(err as Error).message}` }, 500);
-		}
-		return;
-	}
-
-	// GET /api/maintenance/orphaned-index-rows?projectId=...
-	if (url.pathname === "/api/maintenance/orphaned-index-rows" && req.method === "GET") {
-		const projectId = url.searchParams.get("projectId") || undefined;
-		if (!projectId) { json({ error: "Missing projectId" }, 400); return; }
-		const ctx = resolveSearchProject(projectId);
-		if (!ctx) { json({ error: "Project not found" }, 404); return; }
-		await ctx.searchIndex.whenReady();
-		const store = ctx.searchIndex.getStore();
-		if (!store) {
-			json(searchUnavailableResponse(ctx.searchIndex.getState()), 503);
-			return;
-		}
-		try {
-			const rows = store.list({ limit: 100000 });
-			const orphans: Array<{ id: string; source_id: string; parent_id: string | null }> = [];
-			for (const row of rows) {
-				const sourceId = String(row.source_id ?? "");
-				const id = String(row.id ?? "");
-				let isOrphan = false;
-				if (sourceId === "goals") {
-					const goalId = id.replace(/^goal:/, "");
-					isOrphan = !ctx.goalStore.get(goalId);
-				} else if (sourceId === "sessions") {
-					const sessionId = id.replace(/^session:/, "");
-					isOrphan = !ctx.sessionStore.get(sessionId);
-				} else if (sourceId === "messages") {
-					const sessionId = String(row.session_id ?? "");
-					isOrphan = !sessionId || !ctx.sessionStore.get(sessionId);
-				} else if (sourceId === "staff") {
-					const staffId = id.replace(/^staff:/, "");
-					isOrphan = !ctx.staffStore.get(staffId);
-				}
-				if (isOrphan) {
-					orphans.push({
-						id,
-						source_id: sourceId,
-						parent_id: row.parent_id != null ? String(row.parent_id) : null,
-					});
-				}
-			}
-			json({ count: orphans.length, sample: orphans.slice(0, 100) });
-		} catch (err) {
-			json({ error: `Orphan scan failed: ${(err as Error).message}` }, 500);
-		}
-		return;
-	}
-
-	// POST /api/maintenance/cleanup-index-rows
-	if (url.pathname === "/api/maintenance/cleanup-index-rows" && req.method === "POST") {
-		const body = await readBody(req);
-		const projectId = body && typeof body === "object" ? (body as any).projectId : undefined;
-		if (!projectId || typeof projectId !== "string") {
-			json({ error: "Missing projectId" }, 400);
-			return;
-		}
-		const ctx = resolveSearchProject(projectId);
-		if (!ctx) { json({ error: "Project not found" }, 404); return; }
-		await ctx.searchIndex.whenReady();
-		const store = ctx.searchIndex.getStore();
-		if (!store) {
-			json(searchUnavailableResponse(ctx.searchIndex.getState()), 503);
-			return;
-		}
-		try {
-			const rows = store.list({ limit: 100000 });
-			const toDelete: string[] = [];
-			for (const row of rows) {
-				const sourceId = String(row.source_id ?? "");
-				const id = String(row.id ?? "");
-				let isOrphan = false;
-				if (sourceId === "goals") {
-					isOrphan = !ctx.goalStore.get(id.replace(/^goal:/, ""));
-				} else if (sourceId === "sessions") {
-					isOrphan = !ctx.sessionStore.get(id.replace(/^session:/, ""));
-				} else if (sourceId === "messages") {
-					const sessionId = String(row.session_id ?? "");
-					isOrphan = !sessionId || !ctx.sessionStore.get(sessionId);
-				} else if (sourceId === "staff") {
-					isOrphan = !ctx.staffStore.get(id.replace(/^staff:/, ""));
-				}
-				if (isOrphan) toDelete.push(id);
-			}
-			if (toDelete.length) await store.deleteByIds(toDelete);
-			json({ deleted: toDelete.length });
-		} catch (err) {
-			json({ error: `Cleanup failed: ${(err as Error).message}` }, 500);
-		}
-		return;
-	}
-
 	json({ error: "Not found" }, 404);
 }
